@@ -1,8 +1,10 @@
+"""Server for locloc."""
+
 import importlib.resources
 import os
-from datetime import datetime, timedelta, timezone
+from datetime import UTC, datetime, timedelta
 from pathlib import Path
-from typing import Annotated, Optional
+from typing import Annotated
 
 import uvicorn
 from fastapi import FastAPI, HTTPException, Query, Request, Response, status
@@ -42,6 +44,7 @@ app.mount("/static", StaticFiles(directory=resource_root_path / "static_files"))
     status_code=status.HTTP_200_OK,
 )
 def healthcheck() -> Response:
+    """Healthcheck endpoint."""
     return Response(content="OK", media_type="text/plain")
 
 
@@ -51,9 +54,23 @@ async def res(
     request: Request,  # noqa: ARG001
     url: Annotated[HttpUrl, Query(max_length=255)],
     *,
-    branch: Annotated[Optional[str], Query(max_length=255)] = None,  # noqa: FA100
+    branch: Annotated[str | None, Query(max_length=255)] = None,
     is_svg: bool = False,
 ) -> JSONResponse:
+    """Get lines of code statistics for a given repository.
+
+    Args:
+        request (Request): The request object.
+        url (HttpUrl): The URL of the repository.
+        branch (Optional[str]): The branch to analyze. If None, the default branch will be used.
+        is_svg (bool): Whether to return SVG data.
+
+    Returns:
+        JSONResponse: A JSON response containing the lines of code statistics.
+
+    Raises:
+        HTTPException: If there is an error with the git command or if the operation times out.
+    """
     try:
         result, total = get_loc_stats(
             url,
@@ -79,8 +96,21 @@ async def svg(
     request: Request,  # noqa: ARG001
     url: Annotated[HttpUrl, Query(max_length=255)],
     *,
-    branch: Annotated[Optional[str], Query(max_length=255)] = None,  # noqa: FA100
+    branch: Annotated[str | None, Query(max_length=255)] = None,
 ) -> Response:
+    """Get SVG representation of lines of code statistics for a given repository.
+
+    Args:
+        request (Request): The request object.
+        url (HttpUrl): The URL of the repository.
+        branch (Optional[str]): The branch to analyze. If None, the default branch will be used.
+
+    Returns:
+        Response: A response containing the SVG representation of the lines of code statistics.
+
+    Raises:
+        HTTPException: If there is an error with the git command or if the operation times out.
+    """
     try:
         result, _total = get_loc_stats(
             url,
@@ -91,7 +121,7 @@ async def svg(
         raise HTTPException(status_code=status.HTTP_400_BAD_REQUEST) from None
     except TimeoutError:
         raise HTTPException(status_code=status.HTTP_408_REQUEST_TIMEOUT) from None
-    expiry_time = datetime.now(tz=timezone.utc) + timedelta(3666)
+    expiry_time = datetime.now(tz=UTC) + timedelta(3666)
     return Response(
         content=svg,
         media_type="image/svg+xml",
@@ -105,6 +135,14 @@ async def svg(
 
 @app.api_route("/", methods=["GET", "HEAD"], response_class=HTMLResponse)
 async def root(request: Request) -> _TemplateResponse:
+    """Root endpoint that serves the main page.
+
+    Args:
+        request (Request): The request object.
+
+    Returns:
+        _TemplateResponse: A template response containing the main page.
+    """
     return templates.TemplateResponse(
         "index.j2",
         {"request": request, "version": __version__},
@@ -112,6 +150,7 @@ async def root(request: Request) -> _TemplateResponse:
 
 
 def main() -> None:
+    """Main function to run the FastAPI server."""
     config = uvicorn.Config(
         app,
         port=5000,
