@@ -1,4 +1,11 @@
-__version__ = "0.0.5"
+""".. include:: ../README.md"""  # noqa: D415
+
+import importlib.metadata
+
+try:
+    __version__ = importlib.metadata.version(__name__)
+except importlib.metadata.PackageNotFoundError:
+    __version__ = "0.0.0"
 
 from .main import app
 
